Use create-react-app to reformat
